results/rebuttal/run-20231111-034200_LLMatic-zero_shot_donnemartin-system-design-oop-eventplanner_0_20-runs_log/eventplanner/14/LLMatic/admin_dashboard.py
class AdminDashboard:
	def __init__(self):
		self.user_activities = {}
		self.system_performance = {}

	def monitor_user_activities(self, user_id):
		# Mocking user activities
		self.user_activities[user_id] = {'login': '2022-01-01', 'logout': '2022-01-02'}
		return self.user_activities[user_id]

	def manage_user_activities(self, user_id, activity):
		# Mocking managing user activities
		self.user_activities[user_id] = activity
		return self.user_activities[user_id]

	def manage_vendor_listings(self, vendor_id, listing):
		# Mocking managing vendor listings
		self.system_performance[vendor_id] = listing
		return self.system_performance[vendor_id]

	def manage_platform_content(self, content):
		# Mocking managing platform content
		self.system_performance['content'] = content
		return self.system_performance['content']
